Add benchmark for interpreter.
