+++ rust/src/openvasd/controller/entry.rs
@@ -469,12 +469,15 @@ where
 pub mod client {
     use std::sync::Arc;
 
+    use async_trait::async_trait;
     use http::StatusCode;
     use http_body_util::{BodyExt, Empty, Full};
     use hyper::{
         body::Bytes, header::HeaderValue, service::HttpService, HeaderMap, Method, Request,
     };
-    use scannerlib::models::scanner::{self, Scanner};
+    use scannerlib::models::scanner::{
+        self, Error, ScanDeleter, ScanResultFetcher, ScanResults, ScanStarter, ScanStopper, Scanner,
+    };
     use scannerlib::models::{self, Action, Scan, ScanAction, Status};
     use scannerlib::nasl::FSPluginLoader;
     use scannerlib::storage::infisto::{
@@ -490,6 +493,160 @@ pub mod client {
     };
 
     use super::KnownPaths;
+
+    type StartScan = Arc<Box<dyn Fn(Scan) -> Result<(), Error> + Send + Sync + 'static>>;
+    type CanStartScan = Arc<Box<dyn Fn(&Scan) -> bool + Send + Sync + 'static>>;
+    type StopScan = Arc<Box<dyn Fn(&str) -> Result<(), Error> + Send + Sync + 'static>>;
+    type DeleteScan = Arc<Box<dyn Fn(&str) -> Result<(), Error> + Send + Sync + 'static>>;
+    type FetchResults =
+        Arc<Box<dyn Fn(&str) -> Result<ScanResults, Error> + Send + Sync + 'static>>;
+
+    /// A fake implementation of the ScannerStack trait.
+    ///
+    /// This is useful for testing the Scanner implementation.
+    pub struct LambdaScannerBuilder {
+        start_scan: StartScan,
+        can_start_scan: CanStartScan,
+        stop_scan: StopScan,
+        delete_scan: DeleteScan,
+        fetch_results: FetchResults,
+    }
+
+    impl Default for LambdaScannerBuilder {
+        fn default() -> Self {
+            Self::new()
+        }
+    }
+
+    impl LambdaScannerBuilder {
+        pub fn new() -> Self {
+            Self {
+                start_scan: Arc::new(Box::new(|_| Ok(()))),
+                can_start_scan: Arc::new(Box::new(|_| true)),
+                stop_scan: Arc::new(Box::new(|_| Ok(()))),
+                delete_scan: Arc::new(Box::new(|_| Ok(()))),
+                fetch_results: Arc::new(Box::new(|_| Ok(ScanResults::default()))),
+            }
+        }
+
+        pub fn with_start_scan<F>(mut self, f: F) -> Self
+        where
+            F: Fn(Scan) -> Result<(), Error> + Send + Sync + 'static,
+        {
+            self.start_scan = Arc::new(Box::new(f));
+            self
+        }
+
+        pub fn with_can_start_scan<F>(mut self, f: F) -> Self
+        where
+            F: Fn(&Scan) -> bool + Send + Sync + 'static,
+        {
+            self.can_start_scan = Arc::new(Box::new(f));
+            self
+        }
+
+        pub fn with_stop_scan<F>(mut self, f: F) -> Self
+        where
+            F: Fn(&str) -> Result<(), Error> + Send + Sync + 'static,
+        {
+            self.stop_scan = Arc::new(Box::new(f));
+            self
+        }
+
+        pub fn with_delete_scan<F>(mut self, f: F) -> Self
+        where
+            F: Fn(&str) -> Result<(), Error> + Send + Sync + 'static,
+        {
+            self.delete_scan = Arc::new(Box::new(f));
+            self
+        }
+
+        pub fn with_fetch_results<F>(mut self, f: F) -> Self
+        where
+            F: Fn(&str) -> Result<super::ScanResults, Error> + Send + Sync + 'static,
+        {
+            self.fetch_results = Arc::new(Box::new(f));
+            self
+        }
+
+        pub fn build(self) -> LambdaScanner {
+            LambdaScanner {
+                start_scan: self.start_scan,
+                can_start_scan: self.can_start_scan,
+                stop_scan: self.stop_scan,
+                delete_scan: self.delete_scan,
+                fetch_results: self.fetch_results,
+            }
+        }
+    }
+
+    pub struct LambdaScanner {
+        start_scan: StartScan,
+        can_start_scan: CanStartScan,
+        stop_scan: StopScan,
+        delete_scan: DeleteScan,
+        fetch_results: FetchResults,
+    }
+
+    #[async_trait]
+    impl ScanStarter for LambdaScanner {
+        async fn start_scan(&self, scan: Scan) -> Result<(), Error> {
+            let start_scan = self.start_scan.clone();
+            tokio::task::spawn_blocking(move || (start_scan)(scan))
+                .await
+                .unwrap()
+        }
+
+        async fn can_start_scan(&self, scan: &Scan) -> bool {
+            let can_start_scan = self.can_start_scan.clone();
+            let scan = scan.clone();
+            tokio::task::spawn_blocking(move || (can_start_scan)(&scan))
+                .await
+                .unwrap()
+        }
+    }
+
+    #[async_trait]
+    impl ScanStopper for LambdaScanner {
+        async fn stop_scan<I>(&self, id: I) -> Result<(), Error>
+        where
+            I: AsRef<str> + Send + 'static,
+        {
+            let stop_scan = self.stop_scan.clone();
+            let id = id.as_ref().to_string();
+            tokio::task::spawn_blocking(move || (stop_scan)(&id))
+                .await
+                .unwrap()
+        }
+    }
+
+    #[async_trait]
+    impl ScanDeleter for LambdaScanner {
+        async fn delete_scan<I>(&self, id: I) -> Result<(), Error>
+        where
+            I: AsRef<str> + Send + 'static,
+        {
+            let delete_scan = self.delete_scan.clone();
+            let id = id.as_ref().to_string();
+            tokio::task::spawn_blocking(move || (delete_scan)(&id))
+                .await
+                .unwrap()
+        }
+    }
+
+    #[async_trait]
+    impl ScanResultFetcher for LambdaScanner {
+        async fn fetch_results<I>(&self, id: I) -> Result<super::ScanResults, Error>
+        where
+            I: AsRef<str> + Send + 'static,
+        {
+            let fetch_results = self.fetch_results.clone();
+            let id = id.as_ref().to_string();
+            tokio::task::spawn_blocking(move || (fetch_results)(&id))
+                .await
+                .unwrap()
+        }
+    }
 
     type HttpResult = Result<crate::response::Result, scanner::Error>;
     type TypeResult<T> = Result<T, scanner::Error>;
@@ -546,10 +703,9 @@ pub mod client {
         Client::authenticated(scanner, storage)
     }
 
-    pub async fn fails_to_fetch_results() -> Client<
-        scannerlib::scanner::fake::LambdaScanner,
-        Arc<ResultCatcher<inmemory::Storage<crate::crypt::ChaCha20Crypt>>>,
-    > {
+    pub async fn fails_to_fetch_results(
+    ) -> Client<LambdaScanner, Arc<ResultCatcher<inmemory::Storage<crate::crypt::ChaCha20Crypt>>>>
+    {
         use crate::file::tests::example_feeds;
         let storage = crate::storage::inmemory::Storage::default();
         let storage = Arc::new(ResultCatcher::new(storage));
@@ -558,7 +714,7 @@ pub mod client {
             .await
             .unwrap();
 
-        let scanner = scannerlib::scanner::fake::LambdaScannerBuilder::new()
+        let scanner = LambdaScannerBuilder::new()
             .with_fetch_results(|_| Err(scanner::Error::Unexpected("no results".to_string())))
             .build();
         Client::authenticated(scanner, storage)
